% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADNI_disease_stage_bl}
\alias{ADNI_disease_stage_bl}
\title{Predicted disease stage at baseline for ADNI participants}
\format{
A data frame with 2130 rows and 3 variables:
\describe{
  \item{RID}{ADNI roster ID}
  \item{pred_AD_month}{Predicted disease time (in months relative to the average cognitive state of the cognitively normal group)}
}
}
\usage{
ADNI_disease_stage_bl
}
\description{
A dataset containing the predicted disease stage (disease month) for
individuals in ADNI based on their longitudinal 13-item ADAS-cog trajectories
Data are based on a data cut of ADAS-cog scores from July 2019.
}
\keyword{datasets}
