% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progmod.R
\name{GLF}
\alias{GLF}
\title{Overparametrized generalized logistic function.}
\usage{
GLF(t, A, K, B, v, s, c)
}
\arguments{
\item{t}{Time variable}

\item{A}{Lower asymptote parameter}

\item{K}{Upper asymptote parameter}

\item{B}{Time-scaling parameter}

\item{v}{Asymmetry parameter}

\item{s}{Time-shift parameter}

\item{c}{Intercept parameter, should only be used for random effects}
}
\value{
The function values at the supplied time values along with a
        "gradient" attribute.
}
\description{
Overparametrized generalized logistic function.
}
\examples{
GLF(t = c(0, 1), A = 30, K = 0, B = 1, s = 0, v = 1, c = 0)
}
